Refact

+++ sources/Device/src/Menu/MenuItems.h
@@ -99,7 +99,7 @@ struct ButtonOld : public ButtonCommon
 public:
 
     // _highlight - в этом состоянии кнопка находится при первом появлении на экране
-    ButtonOld(pchar _name, pchar _signal, void (*_funcOnPress)(), int _x = -1, int _y = -1) :
+    ButtonOld(pchar _name, pchar _signal, void (*_funcOnPress)(), int16 _x = -1, int16 _y = -1) :
         ButtonCommon(TypeItem::ButtonOld, _name, _name, Font::_1, _x, _y, 0, 0, _funcOnPress),
         name(_name), signal(_signal), x(_x), y(_y)
     {
@@ -135,8 +135,8 @@ private:
 
     int value = 0;
 
-    const int x;
-    const int y;
+    const int16 x;
+    const int16 y;
 
     bool is_visible = true;
